+++ carservice/src/main/resources/com/hubidaauto/carservice/wxapp/core/dao/OrderDao.xml
@@ -269,7 +269,7 @@
                 region_code=#{regionCode,jdbcType=INTEGER},
             </if>
             <if  test="cost!=null">
-                cost=#{cost,jdbcType=DOUBLE},
+                cost=#{cost,jdbcType=INTEGER},
             </if>
         </set>
         where order_id = #{orderId,jdbcType=INTEGER}
