+++ webserver/src/main/java/org/welisdoon/webserver/common/web/CommonAsynService.java
@@ -14,8 +14,8 @@ import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 import org.welisdoon.webserver.common.ApplicationContextProvider;
-import org.welisdoon.webserver.common.JAXBUtils;
-import org.welisdoon.webserver.common.JNIFieldDescriptors;
+import org.welisdoon.common.JAXBUtils;
+import org.welisdoon.common.JNIFieldDescriptors;
 import org.welisdoon.webserver.entity.wechat.messeage.MesseageTypeValue;
 import org.welisdoon.webserver.entity.wechat.messeage.request.RequestMesseageBody;
 import org.welisdoon.webserver.entity.wechat.messeage.response.ResponseMesseage;
